Implemented audio slow motion for the interpolated video
- Fixed GPU exception handler
- Casted a few frame indexes to size_t
- Fixed GPU memory allocation

+++ src/CVObjectDetection.cpp
@@ -83,10 +83,11 @@ void CVObjectDetection::detectObjectsClip(openshot::Clip &video, size_t _start, 
     setProcessingDevice();
 
     size_t frame_number;
+    
     if(!process_interval || end <= 1 || end-start == 0){
         // Get total number of frames in video
-        start = (int)(video.Start() * video.Reader()->info.fps.ToFloat());
-        end = (int)(video.End() * video.Reader()->info.fps.ToFloat());
+        start = (size_t)(video.Start() * video.Reader()->info.fps.ToFloat());
+        end = (size_t)(video.End() * video.Reader()->info.fps.ToFloat());
     }
 
     for (frame_number = start; frame_number <= end; frame_number++)
@@ -128,7 +129,6 @@ void CVObjectDetection::DetectObjects(const cv::Mat &frame, size_t frameId){
 
     // Remove the bounding boxes with low confidence
     postprocess(frame.size(), outs, frameId);
-
 }
