+++ src/CVStabilization.cpp
@@ -67,8 +67,8 @@ void CVStabilization::stabilizeClip(openshot::Clip& video, size_t _start, size_t
     size_t frame_number;
     if(!process_interval || end <= 1 || end-start == 0){
         // Get total number of frames in video
-        start = (int)(video.Start() * video.Reader()->info.fps.ToFloat()) + 1;
-        end = (int)(video.End() * video.Reader()->info.fps.ToFloat()) + 1;
+        start = (size_t)(video.Start() * video.Reader()->info.fps.ToFloat()) + 1;
+        end = (size_t)(video.End() * video.Reader()->info.fps.ToFloat()) + 1;
     }
 
     // Extract and track opticalflow features for each frame
